+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `partial-recall-steal` to take another memory's buffer.
 - Mode key-map.
+- `partial-recall--handle-buffer` now calls new
+  `partial-recall--recollect` instead of `--reclaim`. This will either
+  call `--reclaim` if the buffer is not part of the "reality" or new
+  function `--reinfroce` that will recreate the moment. This makes
+  sure that often-visited buffers aren't forgotten just because they
+  were visited early on.
 
 ### Changed